websocket_client/requirements.txt
websockets>=15.0.1
python-dotenv>=1.1.1